+++ packages/terra-functional-testing/src/config/utils/getConfigurationOptions.js
@@ -10,6 +10,8 @@ const getConfigurationOptions = (options) => {
   const {
     assetServerPort,
     browsers,
+    buildBranch,
+    buildType,
     disableSeleniumService,
     ignoreScreenshotMismatch,
     externalHost,
@@ -34,6 +36,8 @@ const getConfigurationOptions = (options) => {
     hostname: seleniumServiceUrl || gridUrl || (useSeleniumStandaloneService ? 'standalone-chrome' : 'localhost'),
     port: seleniumServicePort || (gridUrl ? 80 : 4444),
     launcherOptions: {
+      buildBranch,
+      buildType,
       disableSeleniumService: disableSeleniumService || useSeleniumStandaloneService || !!gridUrl,
       ignoreScreenshotMismatch,
       formFactor,