+++ packages/terra-functional-testing/tests/jest/config/utils/getConfigurationOptions.test.js
@@ -2,12 +2,15 @@ const path = require('path');
 const getConfigurationOptions = require('../../../../src/config/utils/getConfigurationOptions');
 const getCapabilities = require('../../../../src/config/utils/getCapabilities');
 const getIpAddress = require('../../../../src/config/utils/getIpAddress');
+const { BUILD_BRANCH, BUILD_TYPE } = require('../../../../src/constants/index');
 
 describe('getCapabilities', () => {
   it('should get configuration from cli options', async () => {
     const options = {
       assetServerPort: 8080,
       browsers: ['chrome'],
+      buildBranch: BUILD_BRANCH.pullRequest,
+      buildType: BUILD_TYPE.branchEventCause,
       config: '/path',
       disableSeleniumService: true,
       externalHost: 'externalHost',
@@ -35,6 +38,8 @@ describe('getCapabilities', () => {
       spec: options.spec,
       suite: options.suite,
       launcherOptions: {
+        buildBranch: BUILD_BRANCH.pullRequest,
+        buildType: BUILD_TYPE.branchEventCause,
         disableSeleniumService: true,
         formFactor: options.formFactor,
         gridUrl: options.gridUrl,
@@ -59,6 +64,8 @@ describe('getCapabilities', () => {
     const options = {
       assetServerPort: 8080,
       browsers: ['chrome'],
+      buildBranch: BUILD_BRANCH.master,
+      buildType: BUILD_TYPE.branchEventCause,
       config: '/path',
       disableSeleniumService: false,
       externalHost: 'externalHost',
@@ -85,6 +92,8 @@ describe('getCapabilities', () => {
       spec: options.spec,
       suite: options.suite,
       launcherOptions: {
+        buildBranch: BUILD_BRANCH.master,
+        buildType: BUILD_TYPE.branchEventCause,
         disableSeleniumService: true,
         formFactor: options.formFactor,
         gridUrl: undefined,
@@ -115,6 +124,8 @@ describe('getCapabilities', () => {
       hostname: 'localhost',
       port: 4444,
       launcherOptions: {
+        buildBranch: undefined,
+        buildType: undefined,
         disableSeleniumService: false,
         formFactor: undefined,
         gridUrl: undefined,